efa akaiky mety

+++ inc/minirt.h
@@ -6,7 +6,7 @@
 /*   By: maandria <maandria@student.42antananari    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/01/08 10:40:52 by atolojan          #+#    #+#             */
-/*   Updated: 2025/03/03 17:30:16 by maandria         ###   ########.fr       */
+/*   Updated: 2025/03/05 12:22:09 by maandria         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -109,7 +109,7 @@ t_coord		take_coord_id(t_prog *prog, char *str);
 /************ config *************/
 
 t_viewport	mr_camera_init(float n, t_prog *prog);
-t_coord		mr_pixel_position(t_prog *prog, t_viewport view, int x, int y);
+t_coord		mr_pixel_position(t_prog *prog, t_viewport view, float x, float y);
 
 /************ data *************/
 
@@ -160,7 +160,7 @@ t_coord		op_vector_use(t_coord u, t_coord v);
 t_coord		orientation_ray(t_coord u, t_prog *prog);
 t_coord		op_vector_addition(t_coord u, t_coord v);
 t_coord		op_vpixel_unit(t_viewport view, t_coord u, t_coord v, int index);
-t_coord		op_position_px(t_coord pos_cam,t_viewport view, t_camunit cam, int x, int y);
+t_coord		op_position_px(t_coord pos_cam,t_viewport view, t_camunit cam, float x, float y);
 t_ray		op_quadrique_value_sp(t_coord px,t_prog *prog);
 
 #endif
